Changes to gameboard drawing structure

+++ UI/coordinateconvert.cpp
@@ -6,8 +6,8 @@ namespace Student {
 
 QPointF cubeToAxial(Common::CubeCoordinate cube, int side)
 {
-    int x = side * (sqrt(3) * cube.x  +  sqrt(3)/2 * cube.z);
-    int y = side * (3./2 * cube.z);
+    double x = side * (sqrt(3) * cube.x  +  sqrt(3)/2 * cube.z);
+    double y = side * (3./2 * cube.z);
 
     return QPointF(x, y);
 }
